refactored auth ui components into separate folder and removed redundecy of theme constants
