+++ constants/theme.ts
@@ -87,11 +87,14 @@ export const MindMateColors = {
   deep: "#080C1A",
   surface: "#0C1020",
   lift: "#111827",
+  float: "#1D2845",
   card: "#141C30",
   border: "rgba(148,163,184,0.08)",
   neon: "#A78BFA",
   cyan: "#34D399",
   rose: "#F87171",
+  err: "#F87171",
+  ok: "#86EFAC",
   amber: "#FBBF24",
   lime: "#86EFAC",
   pink: "#F472B6",
@@ -99,8 +102,36 @@ export const MindMateColors = {
   text: "#F8FAFF",
   sub: "#94A3B8",
   muted: "#3D4F6A",
+  glow: "rgba(192,132,252,0.1)",
   a1: "#6D28D9",
   a2: "#0F766E",
   a3: "#9F1239",
   a4: "#1D4ED8",
 };
+
+export const MOOD_DATA = [
+  { e: "😞", l: "Low", c: MindMateColors.rose },
+  { e: "😔", l: "Meh", c: "#F97316" },
+  { e: "😐", l: "Okay", c: MindMateColors.amber },
+  { e: "🙂", l: "Good", c: MindMateColors.lime },
+  { e: "😊", l: "Great", c: MindMateColors.cyan },
+];
+
+export const TAGS_LIST = [
+  "Work 💼",
+  "Anxiety 😰",
+  "Sleep 🌙",
+  "Family 👨‍👩‍👧",
+  "Joy ✨",
+  "Exercise 💪",
+  "Stress 😤",
+  "Gratitude 🙏",
+];
+
+export const INSIGHTS = [
+  "Difficult days are information, not verdicts. What is this feeling trying to tell you?",
+  "You noticed something hard today. That awareness itself is growth.",
+  "Even neutral days hold meaning. What small thing went okay?",
+  "Good energy noticed! What specifically contributed to this feeling?",
+  "You're thriving! Capture what made today great — replicate it.",
+];
